createdSeperateFileForEachComponent
