+++ src/main/Abstract/ECS/Entity/EntityID.hpp
@@ -25,8 +25,10 @@ class EntityID {
 	inline static int IdCounter = 0;
 
 	int getId() const { return this->id; }
-	EntityID(int explicitId) : id(explicitId) {
-		if (explicitId >= IdCounter) IdCounter = explicitId + 1;
+	EntityID(int explicitId) : id(explicitId)
+	{
+		if (explicitId >= IdCounter)
+			IdCounter = explicitId + 1;
 	}
 
   private:
