+++ src/main/Abstract/Overwordl/CollisionSystem.hpp
@@ -1,7 +1,6 @@
 #pragma once
 #include "Abstract/AssetManager/AssetManager.hpp"
 #include "Abstract/ECS/System/System.hpp"
-#include "Components.hpp"
 
 #include <SFML/Graphics/Sprite.hpp>
