+++ src/main/Abstract/Overwordl/Components/ItemComponent.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include "Abstract/ECS/Component/Component.hpp"
+#include "Abstract/TILE_ENUMS.hpp"
+struct ItemComponent : public Component<ItemComponent> {
+	ITEM_TYPE itemType;
+
+	void readFromJson(const nlohmann::json &j) override { itemType = j.value("itemType", ITEM_TYPE()); }
+};
