+++ src/main/Abstract/Overwordl/Components/NPC_COMPONENT.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "Abstract/ECS/Component/Component.hpp"
+
+struct NPC_Component : public Component<NPC_Component> {
+
+	void readFromJson(const nlohmann::json &j) override {}
+};
