+++ src/main/Abstract/Overwordl/Components/PartOfLayerComponent.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "Abstract/ECS/Component/Component.hpp"
+#include "Abstract/TILE_ENUMS.hpp"
+
+struct PartOfLayerComponent : public Component<PartOfLayerComponent> {
+	LEVEL_NAME level;
+	LAYERTYPE layer;
+	void readFromJson(const nlohmann::json &j) override {}
+};
