+++ src/main/Abstract/Overwordl/Components/Player_Component.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "Abstract/ECS/Component/Component.hpp"
+
+struct PlayerComponent : public Component<PlayerComponent> {
+
+	void readFromJson(const nlohmann::json &j) override {}
+};
