+++ src/main/Abstract/Overwordl/Components/RenderComponent.hpp
@@ -1,0 +1,8 @@
+#pragma once
+#include "Abstract/ECS/Component/Component.hpp"
+
+struct RenderComponent : public Component<RenderComponent> {
+
+	int z_layer;
+	void readFromJson(const nlohmann::json &j) override {}
+};
