+++ src/main/Abstract/TILE_ENUMS.hpp
@@ -3,303 +3,303 @@
 #include <unordered_map>
 
 struct TileInfo {
-    int pixelX;
-    int pixelY;
-    int width;
-    int height;
-    };
+	int pixelX;
+	int pixelY;
+	int width;
+	int height;
+};
 
 enum TileType {
-HOUSE_1_RED_00 = 0,
-HOUSE_1_RED_01 = 1,
-HOUSE_1_RED_02 = 2,
-HOUSE_1_RED_03 = 3,
-HOUSE_1_RED_04 = 4,
-HOUSE_1_RED_10 = 5,
-HOUSE_1_RED_11 = 6,
-HOUSE_1_RED_12 = 7,
-HOUSE_1_RED_13 = 8,
-HOUSE_1_RED_14 = 9,
-HOUSE_1_RED_20 = 10,
-HOUSE_1_RED_21 = 11,
-HOUSE_1_RED_22 = 12,
-HOUSE_1_RED_23 = 13,
-HOUSE_1_RED_24 = 14,
-HOUSE_1_RED_30 = 15,
-HOUSE_1_RED_31 = 16,
-HOUSE_1_RED_32 = 17,
-HOUSE_1_RED_33 = 18,
-HOUSE_1_RED_34 = 19,
-HOUSE_1_RED_40 = 20,
-HOUSE_1_RED_41 = 21,
-HOUSE_1_RED_42 = 22,
-HOUSE_1_RED_43 = 23,
-HOUSE_1_RED_44 = 24,
-HOUSE_1_RED_50 = 25,
-HOUSE_1_RED_51 = 26,
-HOUSE_1_RED_52 = 27,
-HOUSE_1_RED_53 = 28,
-HOUSE_1_RED_54 = 29,
-HOUSE_1_BROWN_00 = 30,
-HOUSE_1_BROWN_01 = 31,
-HOUSE_1_BROWN_02 = 32,
-HOUSE_1_BROWN_03 = 33,
-HOUSE_1_BROWN_04 = 34,
-HOUSE_1_BROWN_10 = 35,
-HOUSE_1_BROWN_11 = 36,
-HOUSE_1_BROWN_12 = 37,
-HOUSE_1_BROWN_13 = 38,
-HOUSE_1_BROWN_14 = 39,
-HOUSE_1_BROWN_20 = 40,
-HOUSE_1_BROWN_21 = 41,
-HOUSE_1_BROWN_22 = 42,
-HOUSE_1_BROWN_23 = 43,
-HOUSE_1_BROWN_24 = 44,
-HOUSE_1_BROWN_30 = 45,
-HOUSE_1_BROWN_31 = 46,
-HOUSE_1_BROWN_32 = 47,
-HOUSE_1_BROWN_33 = 48,
-HOUSE_1_BROWN_34 = 49,
-HOUSE_1_BROWN_40 = 50,
-HOUSE_1_BROWN_41 = 51,
-HOUSE_1_BROWN_42 = 52,
-HOUSE_1_BROWN_43 = 53,
-HOUSE_1_BROWN_44 = 54,
-HOUSE_1_BROWN_50 = 55,
-HOUSE_1_BROWN_51 = 56,
-HOUSE_1_BROWN_52 = 57,
-HOUSE_1_BROWN_53 = 58,
-HOUSE_1_BROWN_54 = 59,
-DEFAULT_DEFAULT = 60,
-HOUSE_1_RED = 61,
-SHOTGUN_ = 62,
-HOUSE_1_BROWN = 63,
-GUN_ = 64,
-KNIFE_ = 65,
-HEALINGPACK_IDLE = 66,
-PLAYER_IDLE = 67,
-CONFIG_ = 68,
+	HOUSE_1_RED_00 = 0,
+	HOUSE_1_RED_01 = 1,
+	HOUSE_1_RED_02 = 2,
+	HOUSE_1_RED_03 = 3,
+	HOUSE_1_RED_04 = 4,
+	HOUSE_1_RED_10 = 5,
+	HOUSE_1_RED_11 = 6,
+	HOUSE_1_RED_12 = 7,
+	HOUSE_1_RED_13 = 8,
+	HOUSE_1_RED_14 = 9,
+	HOUSE_1_RED_20 = 10,
+	HOUSE_1_RED_21 = 11,
+	HOUSE_1_RED_22 = 12,
+	HOUSE_1_RED_23 = 13,
+	HOUSE_1_RED_24 = 14,
+	HOUSE_1_RED_30 = 15,
+	HOUSE_1_RED_31 = 16,
+	HOUSE_1_RED_32 = 17,
+	HOUSE_1_RED_33 = 18,
+	HOUSE_1_RED_34 = 19,
+	HOUSE_1_RED_40 = 20,
+	HOUSE_1_RED_41 = 21,
+	HOUSE_1_RED_42 = 22,
+	HOUSE_1_RED_43 = 23,
+	HOUSE_1_RED_44 = 24,
+	HOUSE_1_RED_50 = 25,
+	HOUSE_1_RED_51 = 26,
+	HOUSE_1_RED_52 = 27,
+	HOUSE_1_RED_53 = 28,
+	HOUSE_1_RED_54 = 29,
+	HOUSE_1_BROWN_00 = 30,
+	HOUSE_1_BROWN_01 = 31,
+	HOUSE_1_BROWN_02 = 32,
+	HOUSE_1_BROWN_03 = 33,
+	HOUSE_1_BROWN_04 = 34,
+	HOUSE_1_BROWN_10 = 35,
+	HOUSE_1_BROWN_11 = 36,
+	HOUSE_1_BROWN_12 = 37,
+	HOUSE_1_BROWN_13 = 38,
+	HOUSE_1_BROWN_14 = 39,
+	HOUSE_1_BROWN_20 = 40,
+	HOUSE_1_BROWN_21 = 41,
+	HOUSE_1_BROWN_22 = 42,
+	HOUSE_1_BROWN_23 = 43,
+	HOUSE_1_BROWN_24 = 44,
+	HOUSE_1_BROWN_30 = 45,
+	HOUSE_1_BROWN_31 = 46,
+	HOUSE_1_BROWN_32 = 47,
+	HOUSE_1_BROWN_33 = 48,
+	HOUSE_1_BROWN_34 = 49,
+	HOUSE_1_BROWN_40 = 50,
+	HOUSE_1_BROWN_41 = 51,
+	HOUSE_1_BROWN_42 = 52,
+	HOUSE_1_BROWN_43 = 53,
+	HOUSE_1_BROWN_44 = 54,
+	HOUSE_1_BROWN_50 = 55,
+	HOUSE_1_BROWN_51 = 56,
+	HOUSE_1_BROWN_52 = 57,
+	HOUSE_1_BROWN_53 = 58,
+	HOUSE_1_BROWN_54 = 59,
+	DEFAULT_DEFAULT = 60,
+	HOUSE_1_RED = 61,
+	SHOTGUN_ = 62,
+	HOUSE_1_BROWN = 63,
+	GUN_ = 64,
+	KNIFE_ = 65,
+	HEALINGPACK_IDLE = 66,
+	PLAYER_IDLE = 67,
+	CONFIG_ = 68,
 };
 enum COLLISION_ACTION {
-    KEEP_POSITION = 0,
-    PREV_POSITION = 1,
+	KEEP_POSITION = 0,
+	PREV_POSITION = 1,
 };
 
 enum LAYERTYPE {
-    OVERWORLD = 0,
-    BATTLEWORLD = 1,
+	OVERWORLD = 0,
+	BATTLEWORLD = 1,
 };
 
 enum LEVEL_NAME {
-    LEVEL1 = 0,
-    LEVEL2 = 1,
+	LEVEL1 = 0,
+	LEVEL2 = 1,
 };
 
 enum KEYS {
-    A = 0,
-    B = 1,
-    C = 2,
-    D = 3,
-    E = 4,
-    F = 5,
-    G = 6,
-    H = 7,
-    I = 8,
-    J = 9,
-    K = 10,
-    L = 11,
-    M = 12,
-    N = 13,
-    O = 14,
-    P = 15,
-    Q = 16,
-    R = 17,
-    S = 18,
-    T = 19,
-    U = 20,
-    V = 21,
-    W = 22,
-    X = 23,
-    Y = 24,
-    Z = 25,
-    Num0 = 26,
-    Num1 = 27,
-    Num2 = 28,
-    Num3 = 29,
-    Num4 = 30,
-    Num5 = 31,
-    Num6 = 32,
-    Num7 = 33,
-    Num8 = 34,
-    Num9 = 35,
-    Escape = 36,
-    LControl = 37,
-    LShift = 38,
-    LAlt = 39,
-    LSystem = 40,
-    RControl = 41,
-    RShift = 42,
-    RAlt = 43,
-    RSystem = 44,
-    Menu = 45,
-    LBracket = 46,
-    RBracket = 47,
-    Semicolon = 48,
-    Comma = 49,
-    Period = 50,
-    Apostrophe = 51,
-    Slash = 52,
-    Backslash = 53,
-    Grave = 54,
-    Equal = 55,
-    Hyphen = 56,
-    Space = 57,
-    Enter = 58,
-    Backspace = 59,
-    Tab = 60,
-    PageUp = 61,
-    PageDown = 62,
-    End = 63,
-    Home = 64,
-    Insert = 65,
-    Delete = 66,
-    Add = 67,
-    Subtract = 68,
-    Multiply = 69,
-    Divide = 70,
-    Left = 71,
-    Right = 72,
-    Up = 73,
-    Down = 74,
-    Numpad0 = 75,
-    Numpad1 = 76,
-    Numpad2 = 77,
-    Numpad3 = 78,
-    Numpad4 = 79,
-    Numpad5 = 80,
-    Numpad6 = 81,
-    Numpad7 = 82,
-    Numpad8 = 83,
-    Numpad9 = 84,
-    F1 = 85,
-    F2 = 86,
-    F3 = 87,
-    F4 = 88,
-    F5 = 89,
-    F6 = 90,
-    F7 = 91,
-    F8 = 92,
-    F9 = 93,
-    F10 = 94,
-    F11 = 95,
-    F12 = 96,
-    F13 = 97,
-    F14 = 98,
-    F15 = 99,
-    Pause = 100,
+	A = 0,
+	B = 1,
+	C = 2,
+	D = 3,
+	E = 4,
+	F = 5,
+	G = 6,
+	H = 7,
+	I = 8,
+	J = 9,
+	K = 10,
+	L = 11,
+	M = 12,
+	N = 13,
+	O = 14,
+	P = 15,
+	Q = 16,
+	R = 17,
+	S = 18,
+	T = 19,
+	U = 20,
+	V = 21,
+	W = 22,
+	X = 23,
+	Y = 24,
+	Z = 25,
+	Num0 = 26,
+	Num1 = 27,
+	Num2 = 28,
+	Num3 = 29,
+	Num4 = 30,
+	Num5 = 31,
+	Num6 = 32,
+	Num7 = 33,
+	Num8 = 34,
+	Num9 = 35,
+	Escape = 36,
+	LControl = 37,
+	LShift = 38,
+	LAlt = 39,
+	LSystem = 40,
+	RControl = 41,
+	RShift = 42,
+	RAlt = 43,
+	RSystem = 44,
+	Menu = 45,
+	LBracket = 46,
+	RBracket = 47,
+	Semicolon = 48,
+	Comma = 49,
+	Period = 50,
+	Apostrophe = 51,
+	Slash = 52,
+	Backslash = 53,
+	Grave = 54,
+	Equal = 55,
+	Hyphen = 56,
+	Space = 57,
+	Enter = 58,
+	Backspace = 59,
+	Tab = 60,
+	PageUp = 61,
+	PageDown = 62,
+	End = 63,
+	Home = 64,
+	Insert = 65,
+	Delete = 66,
+	Add = 67,
+	Subtract = 68,
+	Multiply = 69,
+	Divide = 70,
+	Left = 71,
+	Right = 72,
+	Up = 73,
+	Down = 74,
+	Numpad0 = 75,
+	Numpad1 = 76,
+	Numpad2 = 77,
+	Numpad3 = 78,
+	Numpad4 = 79,
+	Numpad5 = 80,
+	Numpad6 = 81,
+	Numpad7 = 82,
+	Numpad8 = 83,
+	Numpad9 = 84,
+	F1 = 85,
+	F2 = 86,
+	F3 = 87,
+	F4 = 88,
+	F5 = 89,
+	F6 = 90,
+	F7 = 91,
+	F8 = 92,
+	F9 = 93,
+	F10 = 94,
+	F11 = 95,
+	F12 = 96,
+	F13 = 97,
+	F14 = 98,
+	F15 = 99,
+	Pause = 100,
 };
 
 enum INTERACTION_TRIGGER {
-    onEnter = 0,
-    byInteractionKey = 1,
+	onEnter = 0,
+	byInteractionKey = 1,
 };
 
 enum INTERACTION_ACTION {
-    START_DIALOG = 0,
-    PICK_ITEM = 1,
-    DOOR_INTERACTION = 2,
-    SWITCH_LAYER = 3,
+	START_DIALOG = 0,
+	PICK_ITEM = 1,
+	DOOR_INTERACTION = 2,
+	SWITCH_LAYER = 3,
 };
 
 enum ITEM_TYPE {
-    WEAPON = 0,
-    HEALING = 1,
-    COLLECTABLE = 2,
+	WEAPON = 0,
+	HEALING = 1,
+	COLLECTABLE = 2,
 };
 
 enum ITEM_WEAPONS_TYPE {
-    MELEEY = 0,
-    DISTANCE = 1,
+	MELEEY = 0,
+	DISTANCE = 1,
 };
 
 enum STATS {
-    STRENGTH = 0,
-    DEXTERITY = 1,
-    FAITH = 2,
+	STRENGTH = 0,
+	DEXTERITY = 1,
+	FAITH = 2,
 };
 
 enum MENUS {
-    Menu1 = 0,
-    Menu2 = 1,
+	Menu1 = 0,
+	Menu2 = 1,
 };
 inline const std::unordered_map<TileType, TileInfo> TILE_DICT = {
-    { HOUSE_1_RED_00, {384, 16, 16, 16}},
-    { HOUSE_1_RED_01, {400, 16, 16, 16}},
-    { HOUSE_1_RED_02, {416, 16, 16, 16}},
-    { HOUSE_1_RED_03, {432, 16, 16, 16}},
-    { HOUSE_1_RED_04, {448, 16, 16, 16}},
-    { HOUSE_1_RED_10, {384, 32, 16, 16}},
-    { HOUSE_1_RED_11, {400, 32, 16, 16}},
-    { HOUSE_1_RED_12, {416, 32, 16, 16}},
-    { HOUSE_1_RED_13, {432, 32, 16, 16}},
-    { HOUSE_1_RED_14, {448, 32, 16, 16}},
-    { HOUSE_1_RED_20, {384, 48, 16, 16}},
-    { HOUSE_1_RED_21, {400, 48, 16, 16}},
-    { HOUSE_1_RED_22, {416, 48, 16, 16}},
-    { HOUSE_1_RED_23, {432, 48, 16, 16}},
-    { HOUSE_1_RED_24, {448, 48, 16, 16}},
-    { HOUSE_1_RED_30, {384, 64, 16, 16}},
-    { HOUSE_1_RED_31, {400, 64, 16, 16}},
-    { HOUSE_1_RED_32, {416, 64, 16, 16}},
-    { HOUSE_1_RED_33, {432, 64, 16, 16}},
-    { HOUSE_1_RED_34, {448, 64, 16, 16}},
-    { HOUSE_1_RED_40, {384, 80, 16, 16}},
-    { HOUSE_1_RED_41, {400, 80, 16, 16}},
-    { HOUSE_1_RED_42, {416, 80, 16, 16}},
-    { HOUSE_1_RED_43, {432, 80, 16, 16}},
-    { HOUSE_1_RED_44, {448, 80, 16, 16}},
-    { HOUSE_1_RED_50, {384, 96, 16, 16}},
-    { HOUSE_1_RED_51, {400, 96, 16, 16}},
-    { HOUSE_1_RED_52, {416, 96, 16, 16}},
-    { HOUSE_1_RED_53, {432, 96, 16, 16}},
-    { HOUSE_1_RED_54, {448, 96, 16, 16}},
-    { HOUSE_1_BROWN_00, {384, 112, 16, 16}},
-    { HOUSE_1_BROWN_01, {400, 112, 16, 16}},
-    { HOUSE_1_BROWN_02, {416, 112, 16, 16}},
-    { HOUSE_1_BROWN_03, {432, 112, 16, 16}},
-    { HOUSE_1_BROWN_04, {448, 112, 16, 16}},
-    { HOUSE_1_BROWN_10, {384, 128, 16, 16}},
-    { HOUSE_1_BROWN_11, {400, 128, 16, 16}},
-    { HOUSE_1_BROWN_12, {416, 128, 16, 16}},
-    { HOUSE_1_BROWN_13, {432, 128, 16, 16}},
-    { HOUSE_1_BROWN_14, {448, 128, 16, 16}},
-    { HOUSE_1_BROWN_20, {384, 144, 16, 16}},
-    { HOUSE_1_BROWN_21, {400, 144, 16, 16}},
-    { HOUSE_1_BROWN_22, {416, 144, 16, 16}},
-    { HOUSE_1_BROWN_23, {432, 144, 16, 16}},
-    { HOUSE_1_BROWN_24, {448, 144, 16, 16}},
-    { HOUSE_1_BROWN_30, {384, 160, 16, 16}},
-    { HOUSE_1_BROWN_31, {400, 160, 16, 16}},
-    { HOUSE_1_BROWN_32, {416, 160, 16, 16}},
-    { HOUSE_1_BROWN_33, {432, 160, 16, 16}},
-    { HOUSE_1_BROWN_34, {448, 160, 16, 16}},
-    { HOUSE_1_BROWN_40, {384, 176, 16, 16}},
-    { HOUSE_1_BROWN_41, {400, 176, 16, 16}},
-    { HOUSE_1_BROWN_42, {416, 176, 16, 16}},
-    { HOUSE_1_BROWN_43, {432, 176, 16, 16}},
-    { HOUSE_1_BROWN_44, {448, 176, 16, 16}},
-    { HOUSE_1_BROWN_50, {384, 192, 16, 16}},
-    { HOUSE_1_BROWN_51, {400, 192, 16, 16}},
-    { HOUSE_1_BROWN_52, {416, 192, 16, 16}},
-    { HOUSE_1_BROWN_53, {432, 192, 16, 16}},
-    { HOUSE_1_BROWN_54, {448, 192, 16, 16}},
-    { DEFAULT_DEFAULT, {0, 0, 16, 16}},
-    { HOUSE_1_RED, {384, 16, 80, 96}},
-    { SHOTGUN_, {16, 112, 16, 16}},
-    { HOUSE_1_BROWN, {384, 112, 80, 96}},
-    { GUN_, {16, 128, 16, 16}},
-    { KNIFE_, {16, 144, 16, 16}},
-    { HEALINGPACK_IDLE, {32, 160, 16, 16}},
-    { PLAYER_IDLE, {464, 160, 16, 16}},
-    { CONFIG_, {16, 176, 16, 16}},
+    {HOUSE_1_RED_00, {384, 16, 16, 16}},
+    {HOUSE_1_RED_01, {400, 16, 16, 16}},
+    {HOUSE_1_RED_02, {416, 16, 16, 16}},
+    {HOUSE_1_RED_03, {432, 16, 16, 16}},
+    {HOUSE_1_RED_04, {448, 16, 16, 16}},
+    {HOUSE_1_RED_10, {384, 32, 16, 16}},
+    {HOUSE_1_RED_11, {400, 32, 16, 16}},
+    {HOUSE_1_RED_12, {416, 32, 16, 16}},
+    {HOUSE_1_RED_13, {432, 32, 16, 16}},
+    {HOUSE_1_RED_14, {448, 32, 16, 16}},
+    {HOUSE_1_RED_20, {384, 48, 16, 16}},
+    {HOUSE_1_RED_21, {400, 48, 16, 16}},
+    {HOUSE_1_RED_22, {416, 48, 16, 16}},
+    {HOUSE_1_RED_23, {432, 48, 16, 16}},
+    {HOUSE_1_RED_24, {448, 48, 16, 16}},
+    {HOUSE_1_RED_30, {384, 64, 16, 16}},
+    {HOUSE_1_RED_31, {400, 64, 16, 16}},
+    {HOUSE_1_RED_32, {416, 64, 16, 16}},
+    {HOUSE_1_RED_33, {432, 64, 16, 16}},
+    {HOUSE_1_RED_34, {448, 64, 16, 16}},
+    {HOUSE_1_RED_40, {384, 80, 16, 16}},
+    {HOUSE_1_RED_41, {400, 80, 16, 16}},
+    {HOUSE_1_RED_42, {416, 80, 16, 16}},
+    {HOUSE_1_RED_43, {432, 80, 16, 16}},
+    {HOUSE_1_RED_44, {448, 80, 16, 16}},
+    {HOUSE_1_RED_50, {384, 96, 16, 16}},
+    {HOUSE_1_RED_51, {400, 96, 16, 16}},
+    {HOUSE_1_RED_52, {416, 96, 16, 16}},
+    {HOUSE_1_RED_53, {432, 96, 16, 16}},
+    {HOUSE_1_RED_54, {448, 96, 16, 16}},
+    {HOUSE_1_BROWN_00, {384, 112, 16, 16}},
+    {HOUSE_1_BROWN_01, {400, 112, 16, 16}},
+    {HOUSE_1_BROWN_02, {416, 112, 16, 16}},
+    {HOUSE_1_BROWN_03, {432, 112, 16, 16}},
+    {HOUSE_1_BROWN_04, {448, 112, 16, 16}},
+    {HOUSE_1_BROWN_10, {384, 128, 16, 16}},
+    {HOUSE_1_BROWN_11, {400, 128, 16, 16}},
+    {HOUSE_1_BROWN_12, {416, 128, 16, 16}},
+    {HOUSE_1_BROWN_13, {432, 128, 16, 16}},
+    {HOUSE_1_BROWN_14, {448, 128, 16, 16}},
+    {HOUSE_1_BROWN_20, {384, 144, 16, 16}},
+    {HOUSE_1_BROWN_21, {400, 144, 16, 16}},
+    {HOUSE_1_BROWN_22, {416, 144, 16, 16}},
+    {HOUSE_1_BROWN_23, {432, 144, 16, 16}},
+    {HOUSE_1_BROWN_24, {448, 144, 16, 16}},
+    {HOUSE_1_BROWN_30, {384, 160, 16, 16}},
+    {HOUSE_1_BROWN_31, {400, 160, 16, 16}},
+    {HOUSE_1_BROWN_32, {416, 160, 16, 16}},
+    {HOUSE_1_BROWN_33, {432, 160, 16, 16}},
+    {HOUSE_1_BROWN_34, {448, 160, 16, 16}},
+    {HOUSE_1_BROWN_40, {384, 176, 16, 16}},
+    {HOUSE_1_BROWN_41, {400, 176, 16, 16}},
+    {HOUSE_1_BROWN_42, {416, 176, 16, 16}},
+    {HOUSE_1_BROWN_43, {432, 176, 16, 16}},
+    {HOUSE_1_BROWN_44, {448, 176, 16, 16}},
+    {HOUSE_1_BROWN_50, {384, 192, 16, 16}},
+    {HOUSE_1_BROWN_51, {400, 192, 16, 16}},
+    {HOUSE_1_BROWN_52, {416, 192, 16, 16}},
+    {HOUSE_1_BROWN_53, {432, 192, 16, 16}},
+    {HOUSE_1_BROWN_54, {448, 192, 16, 16}},
+    {DEFAULT_DEFAULT, {0, 0, 16, 16}},
+    {HOUSE_1_RED, {384, 16, 80, 96}},
+    {SHOTGUN_, {16, 112, 16, 16}},
+    {HOUSE_1_BROWN, {384, 112, 80, 96}},
+    {GUN_, {16, 128, 16, 16}},
+    {KNIFE_, {16, 144, 16, 16}},
+    {HEALINGPACK_IDLE, {32, 160, 16, 16}},
+    {PLAYER_IDLE, {464, 160, 16, 16}},
+    {CONFIG_, {16, 176, 16, 16}},
 };
