+++ src/main/Implementation/main.cpp
@@ -1,6 +1,25 @@
 #include "Abstract/ECS/Component/ComponentRegistry.hpp"
 #include "Abstract/ECS/ECSManager.hpp"
-#include "Abstract/Overwordl/Components.hpp"
+
+#include "Abstract/Overwordl/Components/BoundingBoxComponent.hpp"
+#include "Abstract/Overwordl/Components/CameraComponent.hpp"
+#include "Abstract/Overwordl/Components/CollisionComponent.hpp"
+#include "Abstract/Overwordl/Components/DialogComponent.hpp"
+#include "Abstract/Overwordl/Components/InputComponent.hpp"
+#include "Abstract/Overwordl/Components/InteractionComponent.hpp"
+#include "Abstract/Overwordl/Components/InventoryComponent.hpp"
+#include "Abstract/Overwordl/Components/IsLockedComponent.hpp"
+#include "Abstract/Overwordl/Components/ItemComponent.hpp"
+#include "Abstract/Overwordl/Components/ItemHealstatsComponent.hpp"
+#include "Abstract/Overwordl/Components/ItemWeaponStatsComponent.hpp"
+#include "Abstract/Overwordl/Components/MovementComponent.hpp"
+#include "Abstract/Overwordl/Components/NPC_COMPONENT.hpp"
+#include "Abstract/Overwordl/Components/Player_Component.hpp"
+#include "Abstract/Overwordl/Components/RenderComponent.hpp"
+#include "Abstract/Overwordl/Components/SpriteComponent.hpp"
+#include "Abstract/Overwordl/Components/SwitchLayerComponent.hpp"
+#include "Abstract/Overwordl/Components/TransformComponent.hpp"
+#include "Abstract/Overwordl/Components/WorldComponent.hpp"
 #include "Abstract/Overwordl/WorldParser.hpp"
 
 #include <SFML/Graphics.hpp>
